main menu and race selection big images fix
Former-commit-id: bb83a42d7e0bd4a91f26f2ba21155efec680b562

+++ core/src/com/cauldron/bodyconquest/resourcebars/ResourceBar.java
@@ -17,7 +17,7 @@ public abstract class ResourceBar extends Actor {
     protected Animation<TextureRegion> walkAnimation;
     float stateTime;
     private float elapsedTime;
-    private int insideY;
+    private float insideY;
     private String insideTexturePath;
 
     public ResourceBar(){
@@ -55,11 +55,11 @@ public abstract class ResourceBar extends Actor {
 
     public void setInsideTexturePath(String path){this.insideTexturePath = path;}
 
-    public void setInsideY(int y){
+    public void setInsideY(float y){
         insideY = y;
     }
 
-    private int getInsideY(){
+    private float getInsideY(){
         return insideY;
     }
